Rename client.local to client.base
